Center checkmark in modals (#11016)
Relates #8073 

The checkmark was rendered too-large at 16px because of a global style.
We reuse the same selector and combined with the parent
`positron-modal-dialog-box` class, our custom component's style will
win.

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Center checkmarks in the modal dialogs (e.g. the New Folder from
Template Wizard)

### QA Notes

@:modal @:new-folder-flow

* Reproduce through New -> New Folder from Template or New Folder from
Git
* The checkmark when saving a plot only applies to interactive plots,
e.g. when using the `ggplot-example` in the `qa-example-content` but
also it must have intrinsic size, which I am not sure how to setup so I
just hardcoded to make it appear.