Fix copy/paste in notebooks (#11024)
Addresses #10240

This PR addresses the issue where copying a cell (`C`) and pasting it
inside another cell's editor (`Cmd+V`) would paste the cell metadata
instead of the cell contents.

Additionally, this PR addres support for cross-notebook copy/paste.
Cells can now be copy/pasted between different notebooks which was not
previously supported.

--

Positron notebooks has specific code to handle copy/pasting cell objects
into notebooks. The system clipboard is used to handle pasting into
editors (including cell editors).

When users copied a cell (`C` key or "Copy cell" button), we were
writing the cell json to the notebook clipboard and the system
clipboard. This is why when a users copied a cell (`C` key or "Copy
cell" button) and then pasted inside a cell editor (edit mode), they
would get json instead of the text in the cell.

A code cell with the contents: `print("Hello World")` would return the
following:
```
{
  "cells": [
    {
      "cell_type": "code",
      "source": [
        "print(\"Hello World\")"
      ],
      "metadata": {},
      "outputs": [],
      "execution_count": null
    }
  ],
  "metadata": {
    "kernelspec": {},
    "language_info": {}
  },
  "nbformat": 4,
  "nbformat_minor": 2
}
```
This happened because we were writing cell metadata to the system
clipboard instead of the cell contents. We should have only been writing
the cell contents to the system clipboard.

### Screenshots

**copy/paste single cell into cell editor**


https://github.com/user-attachments/assets/d9b869c1-e8d1-4a9d-a052-b257a5fb6ba2

**copy/paste multiple cells into cell editor**


https://github.com/user-attachments/assets/e262a21b-206f-4c02-a179-56fc23b8d7f7

**copy/paste multiple cells into non-notebook editor**


https://github.com/user-attachments/assets/70f9c22c-ca65-4c4c-898d-1afcefda66f3

**copy/paste cells into a different notebook**


https://github.com/user-attachments/assets/64047f2e-1740-4282-a2a5-e4941dd155fd


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes
@:positron-notebooks

- Copy cell and paste anywhere in a notebook -> pastes new cell with
metadata
- Copy cell and paste within a cell -> paste just the source text
- Copy cell and paste into a non .ipynb file -> paste just the source
text

---------

Signed-off-by: Dhruvi Sompura <dhruvi.sompura@posit.co>
Co-authored-by: Nick Strayer <nick.strayer@posit.co>