Improve notification for explicit "Check for Updates" (#11105)
Addresses #10186

NOTE TO SELF: must revert 79b42994bc4ec64ebf2a4dbeb20d4bda49c1b6c6
before merging

This PR makes changes in the update flow to keep track of whether a
check was explicitly done by a user (as opposed to the behind-the-scenes
automatic checking). It does not align exactly with what is outlined in
https://github.com/posit-dev/positron/issues/10186 as "expected
behavior" because AFAICT that would be _extremely_ difficult with how
the Electron updating/downloading/etc works.

Instead, for the Electron auto-updating that is used on macOS and system
Windows install, what happens here when you manually do "Check for
Updates" is first a notification that there _is_ an update (not sticky,
dismisses itself after a bit):

<img width="1451" height="1114" alt="Screenshot 2025-12-14 at 8 52
50 PM"
src="https://github.com/user-attachments/assets/739315ae-41df-4f4b-9406-d77e734bf8a0"
/>

And then later, a notification that the update is downloaded and you can
restart (sticky):

<img width="1451" height="1114" alt="Screenshot 2025-12-14 at 8 53
04 PM"
src="https://github.com/user-attachments/assets/e4214c72-afe2-4653-91e1-454c63dcc79f"
/>



### Release Notes

#### New Features

- Improved user feedback when "Check for Updates" is manually triggered

#### Bug Fixes

- N/A


### QA Notes

On macOS and Windows, when you explicitly call "Check for Updates..."
either from the gear icon in the lower left or the command palette
_Positron: Check for Updates..._ you should get immediate feedback that
an update either is or is not available. The feedback that it _is_
available is what is new here.

If an update _is_ available, the download will start and then you should
get a notification that you can restart to apply the update once it's
ready.

The behind-the-scenes updating (not triggered by a user) should work the
same as before.

This PR does not involve any changes to what happens on Linux, since
there is no real auto-updating for Linux. We could improve that as well,
but this PR only changes what happens related to the Electron
auto-updating process when triggered by an explicit user action.