Show add cell buttons above and below cells (#11061)
Addresses #10386 and #9923

The "Add Cell" buttons now show up on hover: above the first cell,
between all cells, after the last cell, and when the notebook is empty.
We no longer show the "No Cells" text when a notebook is empty.

**AFTER**


https://github.com/user-attachments/assets/3e835c91-03f6-4816-9af8-e49ab63eb6e3


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

@:positron-notebooks