Use independent console history during debugging (#11102)
Separates 'debug' history entries from others in the history, such that
navigating with <kbd>Up</kbd> and <kbd>Down</kbd> in the Console won't
recall inputs entered during debugging sessions, and vice versa.

The approach is pretty straightforward:
- we store the state of the debugger when recording history entries
- when recalling history entries into the console, we sort them into
'debug' and 'not debug' entries
- when navigating the console history, we use the appropriate history
based on the current debugging state

Addresses https://github.com/posit-dev/positron/issues/4478. 

### Release Notes

#### New Features

- Use a separate console history during debugging sessions (#4478) 

#### Bug Fixes

- N/A


### QA Notes

This approach keeps things simple by not requiring any changes to the
kernels (which currently do not explicitly tell the UI whether they are
debugging); it just uses the _global_ debugging session, i.e. "is the UI
in debug mode". This creates a new (but smaller) problem: if you have
multiple sessions, the sessions that are _not_ currently in debug mode
are still going to be using their alternate (debug) histories. This is
probably not going to be common in practice, but if it is, we could make
a more nuanced fix here.

---------

Signed-off-by: Jonathan <jonathan@posit.co>
Co-authored-by: Jennifer (Jenny) Bryan <jenny.f.bryan@gmail.com>