[python] look for web app creation before imports  (#11083)
Currently, we just look at `import XXX` or `from XXXX import XXXX` to
decide if users are in a certain type of web app. This isn't the most
reliable. This PR adds in a preliminary check of "has a user created the
actual app object from a framework" rather than just relying on imports,
which we keep as a backup.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- #10258 Adjusted web app detection to look for app creation before
imports.


### QA Notes

#10258 has a good example, and I've added a few others in unittests.
Generally, creating apps with multiple imports and making sure we detect
the one that is initialized by, eg, `app = Dash()` or
`st.write("hello")`