Discriminate log messages by session ID (#10961)
Branched from https://github.com/posit-dev/positron/pull/10933
Progress towards https://github.com/posit-dev/positron/issues/10121
Addresses https://github.com/posit-dev/positron/issues/7460

This PR:

- Makes the kernel logs less confusing when multiple sessions are
running concurrently. This should help making sense of CI logs for
#10121.

- Reduces the number of output channels, making it easier for users to
find and report the relevant channels.


Approach:

- There is now only one `Kernel` channel, one `Supervisor` channel, and
one `Language Server` channel for _all_ sessions, console or notebook.
- All log messages are prefixed with the session ID (which includes a
`notebook` particle for notebooks).
- The language server does not need the session prefix because there is
only one active at a time, so the session start message, which includes
the session ID, is sufficient to disambiguate sessions.

This is going to make logs more busy but will greatly improve
disambiguation in debugging session. The "Filter" feature of output
channels can also be used to filter messages for a particular session,
if needed.

The alternative approach I first considered was to include the session
ID in output channel names. The issue is that this leads to an explosion
of output channels (3 by session ID) that are never cleaned up.

If we go forward with this, I'll send a companion PR to update
https://positron.posit.co/troubleshooting.html#python-and-r-logs.

Here is how this looks:


https://github.com/user-attachments/assets/cb8f8ece-1172-4b20-8f86-6c72b468651a



https://github.com/user-attachments/assets/d2bdd03d-78ca-4a5b-92df-785ae00438e1



https://github.com/user-attachments/assets/811ccd07-9304-47cc-9122-65a8b8dbb00d



<!-- Thank you for submitting a pull request.
If this is your first pull request you can find information about
contributing here:
  * https://github.com/posit-dev/positron/blob/main/CONTRIBUTING.md

We recommend synchronizing your branch with the latest changes in the
main branch by either pulling or rebasing.
-->

<!--
  Describe briefly what problem this pull request resolves, or what
  new feature it introduces. Include screenshots of any new or altered
  UI. Link to any GitHub issues but avoid "magic" keywords that will
  automatically close the issue. If there are any details about your
  approach that are unintuitive or you want to draw attention to, please
  describe them here.
-->


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Log output channels have been consolidated. There is now a single
`Kernel`, `Supervisor`, and `Language Server` channel per language.

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->