Postpone executing `props.actions()` in the ActionBarMenuButton
We want to execute the `props.actions` function as late as possible to
ensure we have the most up-to-date data.

Previously, this component would execute the `prop.actions` function
when the compenent was first rendered. It would then memoize the list of
actions in a `React.useState`. This led to stale state, most especially
with regard to the `enabled` property.

If we compute the list of actions at render-time, we have stronger
guarantees that the rendered data is accurate.

This does not take into consideration the possibility that the command
is disabled while the menu is visible, but that is unlikely an area of
concern.

Relates #8728