Don't create a Variables backend when pane is hidden (#11133)
This change makes it possible to hide the Variables view.

<img width="470" height="226" alt="image"
src="https://github.com/user-attachments/assets/962e056c-7563-4a49-a562-750474535697"
/>

It also causes the backends (variables comms) to track the view state,
such that they are closed (or never opened) if the view is hidden.

Taken together, these changes make it possible to prevent any possible
interference from the variables comms when working in Positron, without
adding any additional UX or settings.

Addresses https://github.com/posit-dev/positron/issues/507.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- The Variables view can now be hidden, for troubleshooting or when
working with large or sensitive objects in memory. (#507)

#### Bug Fixes

- N/A


### QA Notes

You can use the debug Runtime Sessions view (enable in Settings) to
verify that the variables comm is closed when the pane is hidden, and
recreated when it is opened. You want to see something like this, which
shows an old variables comm (prior to hiding) and a new one (opened
after showing again).

<img width="540" height="418" alt="image"
src="https://github.com/user-attachments/assets/cd71583a-5c42-4eae-b4f6-9c0ecc4af067"
/>

Only hiding the view entirely (via the hide command) should close the
comm; obscuring or collapsing the view shouldn't do anything.

Finally, note that when the variables pane is hidden from the user, it's
also hidden from LLMs, since the inspection tools mostly work by using
the variables comm. This could be considered a feature, or a bug,
depending on your perspective!