Only register data explorer copy keybindings with focus (#11074)
The copy keybindings should only be registered when the data explorer
has focus. Without this, the data explorer will copy the currently
selected cell to the clipboard even if focus is entirely elsewhere in
Positron (e.g. the Positron Assistant chat input).

Relates #10669

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Only copy data explorer contents if the data explorer has focus


### QA Notes

Open the data explorer and select a cell. If you then focus on another
pane, e.g. Positron Assistant, and attempt to copy anything such as your
own chat input -- the selected cell in the data explorer will be copied
to the clipboard, not the chat input.

Note that this only affects keybindings. The context menu (which also
supports copy) is out of scope.

@:data-explorer