Positron Notebooks: Find (#11056)
Addresses #10064.

I added a notebook editor extension system based on the upstream ones
for Monaco and Notebooks. Looked like a useful way to separate out
components like this.

There were a bunch of existing find widget implementations, but
unfortunately they all differ in subtle ways that made them unusable in
our case. I made a generic React-based find widget that we can use in
other places too like the Console. The idea is to also extend that
widget with replace functionality when we get there.

I started looking into unit/integration tests but it got complicated so
will do that separately.

_Note: I used Claude pretty heavily for this PR, but I kept a heavy hand
on the architecture throughout._

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- N/A

### QA Notes

@:positron-notebooks @:web

---------

Co-authored-by: Marie Idleman <marie.idleman@posit.co>
Co-authored-by: Nick Strayer <nick.strayer@posit.co>